Splitting the hero section from main page
